requirements.txt
boto3==1.21.35
botocore==1.24.35
Flask==1.1.4
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
MarkupSafe==2.0.1
python-dotenv==0.20.0
stripe==2.62.0
Werkzeug==1.0.1